+++ src/core/common/swagger/index.js
@@ -1,3 +1,1 @@
 export * from './filter';
-export * from './objectId';
-export * from './upload-file';